fastapi/01_fastapi_tutorial/requirements.txt
fastapi==0.109.2
pydantic==2.6.1
uvicorn[standard]==0.27.1
pydantic[email]==2.1.0.post1
python-multipart==0.0.9
python-jose[cryptography]=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.27
orjson==3.9.15